move client example to single_threaded example
