+++ examples/single_threaded.rs
@@ -1,5 +1,6 @@
 #![deny(warnings)]
 
+use http_body_util::BodyExt;
 use hyper::server::conn::http2;
 use std::cell::Cell;
 use std::net::SocketAddr;
@@ -8,10 +9,13 @@ use tokio::net::TcpListener;
 
 use hyper::body::{Body as HttpBody, Bytes, Frame};
 use hyper::service::service_fn;
+use hyper::Request;
 use hyper::{Error, Response};
 use std::marker::PhantomData;
 use std::pin::Pin;
 use std::task::{Context, Poll};
+use std::thread;
+use tokio::net::TcpStream;
 
 struct Body {
     // Our Body type is !Send and !Sync:
@@ -40,23 +44,44 @@ impl HttpBody for Body {
     }
 }
 
-fn main() -> Result<(), Box<dyn std::error::Error>> {
+fn main() {
     pretty_env_logger::init();
 
-    // Configure a runtime that runs everything on the current thread
-    let rt = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .expect("build runtime");
+    let server = thread::spawn(move || {
+        // Configure a runtime for the server that runs everything on the current thread
+        let rt = tokio::runtime::Builder::new_current_thread()
+            .enable_all()
+            .build()
+            .expect("build runtime");
 
-    // Combine it with a `LocalSet,  which means it can spawn !Send futures...
-    let local = tokio::task::LocalSet::new();
-    local.block_on(&rt, run())
+        // Combine it with a `LocalSet,  which means it can spawn !Send futures...
+        let local = tokio::task::LocalSet::new();
+        local.block_on(&rt, server()).unwrap();
+    });
+
+    let client = thread::spawn(move || {
+        // Configure a runtime for the client that runs everything on the current thread
+        let rt = tokio::runtime::Builder::new_current_thread()
+            .enable_all()
+            .build()
+            .expect("build runtime");
+
+        // Combine it with a `LocalSet,  which means it can spawn !Send futures...
+        let local = tokio::task::LocalSet::new();
+        local
+            .block_on(
+                &rt,
+                client("http://localhost:3000".parse::<hyper::Uri>().unwrap()),
+            )
+            .unwrap();
+    });
+
+    server.join().unwrap();
+    client.join().unwrap();
 }
 
-async fn run() -> Result<(), Box<dyn std::error::Error>> {
+async fn server() -> Result<(), Box<dyn std::error::Error>> {
     let addr: SocketAddr = ([127, 0, 0, 1], 3000).into();
-
     // Using a !Send request counter is fine on 1 thread...
     let counter = Rc::new(Cell::new(0));
 
@@ -84,6 +109,45 @@ async fn run() -> Result<(), Box<dyn std::error::Error>> {
             }
         });
     }
+}
+
+async fn client(url: hyper::Uri) -> Result<(), Box<dyn std::error::Error>> {
+    let host = url.host().expect("uri has no host");
+    let port = url.port_u16().unwrap_or(80);
+    let addr = format!("{}:{}", host, port);
+    let stream = TcpStream::connect(addr).await?;
+
+    let (mut sender, conn) = hyper::client::conn::http2::handshake(LocalExec, stream).await?;
+
+    tokio::task::spawn_local(async move {
+        if let Err(err) = conn.await {
+            println!("Connection failed: {:?}", err);
+        }
+    });
+
+    let authority = url.authority().unwrap().clone();
+
+    // Make 4 requests
+    for _ in 0..4 {
+        let req = Request::builder()
+            .uri(url.clone())
+            .header(hyper::header::HOST, authority.as_str())
+            .body(Body::from("test".to_string()))?;
+
+        let mut res = sender.send_request(req).await?;
+
+        println!("Response: {}", res.status());
+        println!("Headers: {:#?}\n", res.headers());
+
+        // Print the response body
+        while let Some(next) = res.frame().await {
+            let frame = next?;
+            if let Some(chunk) = frame.data_ref() {
+                println!("data {:#?}", chunk)
+            }
+        }
+    }
+    Ok(())
 }
 
 // NOTE: This part is only needed for HTTP/2. HTTP/1 doesn't need an executor.
